feat(#13): transfere tipagem pro arquivo index.ts
