+++ src/components/SessionDetails/SessionActions/SessionActions.tsx
@@ -3,12 +3,12 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faFilter } from "@fortawesome/free-solid-svg-icons";
 import styles from "./SessionActions.module.scss";
 import SessionActionCard from "../SessionActionCard/SessionActionCard.tsx";
-import Select from "../../elements/Select";
 import type {
   ActionTypeFilter,
   SessionAnalysisSection,
   ViewMode,
 } from "../../../pages/SessionView";
+import Select from "../../elements/Select/Select.tsx";
 
 type Props = {
   viewMode: ViewMode;
