+++ src/components/elements/Select/Select.tsx
@@ -1,24 +1,6 @@
-import type { ChangeEvent, SelectHTMLAttributes } from "react";
 import styles from "./Select.module.scss";
-
-type SelectValue = string | number;
-
-export type SelectOption<TValue extends SelectValue = string> = {
-  label: string;
-  value: TValue;
-  disabled?: boolean;
-};
-
-export type SelectProps<TValue extends SelectValue = string> = {
-  value: TValue | "";
-  options: SelectOption<TValue>[];
-  onChange: (value: TValue | "", event: ChangeEvent<HTMLSelectElement>) => void;
-  placeholder?: string;
-  label?: string;
-  disabled?: boolean;
-  className?: string;
-  name?: string;
-} & Omit<SelectHTMLAttributes<HTMLSelectElement>, "value" | "onChange" | "disabled" | "className" | "name">;
+import type { ChangeEvent } from "react";
+import type { SelectValue, SelectProps } from './index.ts';
 
 const Select = <TValue extends SelectValue = string>({
   value,
